fixed warning for multiple classes in one file
